+++ sartd/src/fib.rs
@@ -1,9 +1,10 @@
 pub(crate) mod api_server;
-pub(crate) mod route;
-pub(crate) mod server;
-pub(crate) mod error;
-pub(crate) mod config;
-pub(crate) mod channel;
-pub(crate) mod kernel;
 pub(crate) mod bgp;
+pub(crate) mod channel;
+pub(crate) mod config;
+pub(crate) mod error;
+pub(crate) mod kernel;
 pub(crate) mod rib;
+pub(crate) mod route;
+pub(crate) mod rt_client;
+pub(crate) mod server;
